+++ src/main/java/com/example/springboot/controllers/UserController.java
@@ -63,8 +63,8 @@ public class UserController {
     @GetMapping("/search")
     @Secured({PermissionCode.SYSTEM_ADMIN})
     public ResponseEntity<CommonResponse<PaginationResponseData<UserDTO>>> search(
-            @RequestParam(required = false) Integer pageNumber,
-            @RequestParam(required = false) Integer pageSize,
+            @RequestParam(required = false) Long pageNumber,
+            @RequestParam(required = false) Long pageSize,
             @RequestParam(required = false) String username,
             @RequestParam(required = false) String email,
             @RequestParam(required = false) String status,
